requirements.txt
streamlit
matplotlib
Pillow
numpy